Fix `collection_ops_test` for Databricks 14.3
Fixes #11532.

This commit introduces a RapidsErrorUtils shim for Databricks 14.3, to
deal with the new error messages thrown for large array/sequences.

This should fix the failure in `collection_ops_test.py::test_sequence_too_long_sequence`
on Databricks 14.3.
